Tagging as v0.20  Added in support for AlarmState to be toggled when relays are turned on and off.  Creates new issue #19.  Removed MenuTimeSet, and MenuYesNo from the menu system as they are no longer needed in favor of MenuNumSel.  Adjusted menu delay from 100ms to 200ms to see if that helps with the menu exitin issue #17.

+++ Alarms.ino
@@ -242,7 +242,7 @@ void AlarmSet(byte id)
 
 	rd = MenuNumSel(4, 255, start, 0, 1, 1, 8, 1, 250);	//	Call the function to edit the variable
 
-	int bit;
+	byte bit;
 	bit = 1 << id;
 
 	if ((serialDebug & 4) == 4){ Serial.print("AlarmEnable: "); }
